% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFiles.R
\name{splitFiles}
\alias{splitFiles}
\title{Split the \pkg{kana} export file into its components}
\usage{
splitFiles(path, state.path = "state.h5", file.dir = "files")
}
\arguments{
\item{path}{String containing the path to the \pkg{kana} export file.}

\item{state.path}{String containing the path to the output HDF5 file for the analysis state and parameters.}

\item{file.dir}{String containing the path to the output directory for the embedded input files.}
}
\value{
The analysis state is saved to \code{state.path}.
For exports containing embedded inputs, \code{file.dir} is created and filled with the separated input files.
A list is invisibly returned containing \code{type}, whether the export contained linked or embedded files;
and \code{version}, a version number for the exported file.
}
\description{
Split the analysis file exported by the \pkg{kana} application into its components.
Specifically, this refers to a file for the analysis parameters and results,
and then further embedded files corresponding to the original inputs.
}
\details{
Files in \code{file.dir} are named by their index in the analysis parameters of the \code{inputs} step.
Ideally we would give them a more intuitive name; however, there is no guarantee that the input files supplied to \pkg{kana} have unique names.
}
